Add support for trusted-firmware-a releases
The Makefile for the trusted-firmware-a software is a bit special since
it has no EXTAVERSION variable instead they use a VERSION_STRING
variable. The variable is composed by:

| # Default build string (git branch and commit)
| ifeq (${BUILD_STRING},)
|         BUILD_STRING  :=  $(shell git describe --always --dirty --tags 2> /dev/null)
| endif
| VERSION_STRING    :=  v${VERSION_MAJOR}.${VERSION_MINOR}(${BUILD_TYPE}):${BUILD_STRING}

Append the final VERSION_STRING variable with the umpf version to be
independent of the set/not set BUILD_STRING variable.

The output before this patch was (BUILD_STRING not set):

| NOTICE:  BL31: v2.7(release):v2.7

With this patch applied the output will be:

| NOTICE:  BL31: v2.7(release):v2.7-20221109-2

Signed-off-by: Marco Felsch <m.felsch@pengutronix.de>