umpf: add branch choice in case of additional commits
When umpf compares branches between all of the known remotes, it warns
when a discovered branch includes additional commits compared to the
one on the selected remote. The user is warned, but has no choice to
use the alternative remote and include the commits.

This commit adds a choice to the user in the situation described above.
This is useful to facilitate generating an umpf-tag with some local
changes when working on projects with multiple developers pushing to the
same common branches, i.e. where the remote is not to be updated until
the changes are verified and ready to merge.

Signed-off-by: Emil-Juhl <emdj@bang-olufsen.dk>