umpf: add new --override option to override topic commit-ish
In one project, umpf tags are pushed immediately, but branches are only
pushed upstream once they are reviewed and cleared for merge into BSP.
It's thus often the case, that developers want to use the upstream
branches, except for that one branch they are currently working on.

This is not an unusual case and the tips-and-tricks lists some
workarounds to achieve this. Provide a better way by allowing overriding
single topic branches on the command-line.

Signed-off-by: Ahmad Fatoum <a.fatoum@pengutronix.de>