umpf: call rev-parse on manually selected local branches as well
When umpf compares branches between all of the known remotes, it will now
prompt the user when it discovers a branch that includes additional commits
compared to the one on the selected remote.

If the user confirms, the branch name is taken as is, which would result in
an umpf-hashinfo of e.g. heads/v6.4/topic/something, which breaks the
ability of doing identical umpfs.

To restore that ability, we use the result of calling rev-parse on the
${branch}, which is conveniently available already in ${b}.

Fixes: 000e642bc702 ("umpf: add branch choice in case of additional commits")
Signed-off-by: Ahmad Fatoum <a.fatoum@pengutronix.de>