README: Formalize signing off commits
The automatic checks for pull requests on Github include a check that
fails if a Signed-off-by: line by the commiter is missing in any commit.

For that to make sense formalize the meaning of such a line. I'm not a
lawyer (thank heavens!) but I believe that without such a formalisation
you cannot safely assume the "usual semantic also used in other
open-source projects".

Signed-off-by: Uwe Kleine-König <u.kleine-koenig@pengutronix.de>