umpf: fix losing flags while continue
The 'umpf continue' case was tested only with the use-case where flags
are specified via the cmdline. If flags are provided via an useries file
they are still lost. Fix this by making use of the new flags status
file while bailout(). This file is always generated.

This commit also fixes the case the following case: flags were succefull
verified but umpf fails later on e.g. during a tag run and we would need
to run 'umpf continue'. In such case they are marked within the
${STATE}/done file so the hooks are not executed again. Fix this by
checking for the FLAGS cmdline parameter or the new ${STATE}/flags file
to parse and setup the flags everytime.

Signed-off-by: Marco Felsch <m.felsch@pengutronix.de>