format-patch: correctly set UMPF_VERSION with --bb
When format-patch'ing a series for bitbake, UMPF_VERSION is taken from
${STATE}/version, but this file is never written when we don't set
STABLE for the "version" branch of parse(). Additionally, the version
must be parsed back from the input series in that case, otherwise it
will always be initialised to "-1", with the result that UMPF_VERSION
always ends in "-1" even when format-patch'ing an umpf that was built
with '-v2' or greater.

Update the test case with an appropriate tag name.

Fixes: 8d0d4ccac5b899526e08 (2024-03-06, "umpf: output version infos to bitbake variables during format-patch")
Signed-off-by: Roland Hieber <rhi@pengutronix.de>