umpf: do not set PV in series.inc
Since most recipes tend to use PV in their SRC_URI and we cannot set it
as a default value ('?='), it is the safest way to not set it
automatically from series.inc.

Users who want to have the umpf version as part of the yocto package
version can set

  PV = "${UMPF_PV}"

at an appropriate location within the recipe.

Signed-off-by: Enrico Jörns <ejo@pengutronix.de>