umpf: answer interactive prompts with default when --default set
When building or tagging an umpf, umpf can ask which branch to use when
multiple are available. Piping yes(1) into umpf in this case doesn't
work, because the variable populated by read in that case is left empty.

Provide for this use case, a --default option that will take a default
if one is known.

Signed-off-by: Ahmad Fatoum <a.fatoum@pengutronix.de>