umpf: make LINUX_VERSION a weak assignment
Umpf already supports stable updates outside of umpf if the
extraversion=conflictfree was provided for the umpf-tag.

This patch allow the Yocto users to override LINUX_VERSION from anywhere
within a recipe which can be the case if the initial umpf-tag was
provided by another team than the one doing the maintenance work (stable
updates).

Signed-off-by: Marco Felsch <m.felsch@pengutronix.de>