umpf: keep the first commit-ish for a branch when creating the umpf from an umerge
When an umpf is created from a umerge, the same topic branch may be
found multiple times:
1. The actual merge for the topic in the umerge
2. as a base for another topic

The commit-ish from the first case is the relevant one. All others may
use older versions of the topic branch, so ignore those when creating
the useries.

Without this, creating an identical tag from a umerge will use the wrong
version of the topic branch.

Signed-off-by: Michael Olbrich <m.olbrich@pengutronix.de>