umpf: preserve flags while creating a umerge
The flags are never stored in a file. So creating the umpf-build-flags
note from a file is a noop and the flags are lost.

Instead, handle flags like the 'relative'. Set the FLAGS variable when
flags are found and use that to create the note.

This is a modified version of #60 from Alvin Šipraga <alvin@pqrs.dk>.

Signed-off-by: Michael Olbrich <m.olbrich@pengutronix.de>