umpf: display flags when showing a umerge
When runing 'umpf show' on an umerge, all lines are explicitly
generated. There is no code for the flags, so they are not shown even
though they are available internally and running e.g. 'umpf tag'
propagates the flags correctly.

Add the relevant code so that 'umpf show' works correctly as well.

Signed-off-by: Michael Olbrich <m.olbrich@pengutronix.de>